requirements.txt
django
djangorestframework
psycopg2
python-decouple
drf-yasg
